refactor(vibe3): break circular dependency between utils and services
Move get_manager_usernames() and get_handoff_state_label() from
services/orchestra_helpers.py to config/manager_config.py to eliminate
utils ↔ services circular dependency.

This is Priority 1 fix in the circular dependency elimination roadmap.

Changes:
- Create src/vibe3/config/manager_config.py with moved functions
- Update src/vibe3/utils/comment_utils.py import path
- Add backward compatibility re-exports in orchestra_helpers.py
- ConventionResolver already in config module (no move needed)

Impact:
- ✅ Breaks utils ↔ services circular dependency
- ✅ Reduces total circular dependencies from 101 → 64 (-37%)
- ✅ utils becomes pure utility module (no business dependencies)
- ✅ Aligns with layered architecture: config → utils → services

Test results:
- ✅ All utils tests pass (99 passed)
- ✅ Function behavior unchanged (backward compatible)
- ✅ Import chain validated

Related: #1966
